+++ include/objectbox-sync.hpp
@@ -21,7 +21,7 @@
 #include "objectbox-sync.h"
 #include "objectbox.hpp"
 
-static_assert(OBX_VERSION_MAJOR == 4 && OBX_VERSION_MINOR == 3 && OBX_VERSION_PATCH == 1,  // NOLINT
+static_assert(OBX_VERSION_MAJOR == 5 && OBX_VERSION_MINOR == 0 && OBX_VERSION_PATCH == 0,  // NOLINT
               "Versions of objectbox.h and objectbox-sync.hpp files do not match, please update");
 
 namespace obx {
@@ -92,7 +92,6 @@ public:
     static SyncCredentials jwtCustomToken(const std::string& token) {
         return SyncCredentials(OBXSyncCredentialsType::OBXSyncCredentialsType_JWT_CUSTOM, token);
     }
-
 };
 
 /// Listens to login events on a sync client.
@@ -397,6 +396,40 @@ public:
         }
         internal::checkErrOrThrow(err);
     }
+
+    /// Adds or replaces a sync filter variable value to the sync client.
+    /// Client filter variables can be used in server-side sync filters to filter out objects that do not match the
+    /// filter. Filter variables must be added before login, e.g. before obx_sync_start() or setting credentials.
+    /// @param name non-NULL name of the filter variable
+    /// @param value non-NULL value of the filter variable
+    void putFilterVariable(const char* name, const char* value) {
+        obx_err err = obx_sync_filter_variables_put(cPtr(), name, value);
+        internal::checkErrOrThrow(err);
+    }
+
+    /// Adds or replaces a sync filter variable value to the sync client.
+    /// Client filter variables can be used in server-side sync filters to filter out objects that do not match the
+    /// filter. Filter variables must be added before login, e.g. before obx_sync_start() or setting credentials.
+    /// @param name name of the filter variable
+    /// @param value value of the filter variable
+    void putFilterVariable(const std::string& name, const std::string& value) {
+        putFilterVariable(name.c_str(), value.c_str());
+    }
+
+    /// Removes all previously added sync filter variable values.
+    void removeAllFilterVariables() {
+        obx_err err = obx_sync_filter_variables_remove_all(cPtr());
+        internal::checkErrOrThrow(err);
+    }
+
+    /// Removes a previously added sync filter variable value.
+    void removeFilterVariable(const char* name) {
+        obx_err err = obx_sync_filter_variables_remove_all(cPtr());
+        internal::checkErrOrThrow(err);
+    }
+
+    /// Removes a previously added sync filter variable value.
+    void removeFilterVariable(const std::string& name) { removeFilterVariable(name.c_str()); }
 
     /// Triggers a reconnection attempt immediately.
     /// By default, an increasing backoff interval is used for reconnection attempts.