+++ doxygen/Changelog.md
@@ -3,6 +3,29 @@
 ObjectBox C and C++ API Changelog
 =================================
 
+5.0.0-rc (2025-09-30)
+------------------
+
+### User-Specific Data Sync
+
+* Sync filters: define server-side filter expression to sync individual data for each sync user.
+  This is also known as "user-specific data sync" and requires Sync clients version 5.0.
+* Client variables: clients may define key/value pairs that can be used in sync filters
+
+### Fixes
+
+* Fixed clearing 1:N backlinks for IDs larger than 32-bit (setting backlink ID to 0 on the "1" side)
+* In-memory with WAL file: improved error handling
+* Safeguard against undefined behavior by panicking in rare illegal usage patterns that are not recoverable.
+  I.e. deleting a (write) transaction in a non-owner thread cannot be safely handled in any other way.
+
+### Examples
+
+* Make each example self-contained, you can e.g. copy an example's directory as a starting point for your own app
+* Add a convenient `build.sh` script to each example that works the same way across examples,
+  e.g. `./build.sh run` to build and run the example in one step
+* Make sources more readable (refactorings, added additional comments)
+
 4.3.1 (2025-07-28)
 ------------------
 * Cursor/Query: deleting a cursor (e.g. in a non-creator thread) waits for any query to finish